+++ pkg/util/k8sutil/services.go
@@ -30,6 +30,7 @@ import (
 	core "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/apis/shared"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/service"
@@ -145,7 +146,7 @@ func CreateDatabaseClientService(ctx context.Context, svcs servicev1.ModInterfac
 	svcName := CreateDatabaseClientServiceName(deploymentName)
 	ports := []core.ServicePort{
 		{
-			Name:     "server",
+			Name:     api.ServerGroupReservedContainerNameServer,
 			Protocol: core.ProtocolTCP,
 			Port:     shared.ArangoPort,
 		},