+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - (Bugfix) Proper agent cleanout
 - (Bugfix) Fix ClusterScaling integration
 - (Feature) Sensitive information protection
+- (Bugfix) Propagate SecurityContext to the ID Containers
 
 ## [1.2.15](https://github.com/arangodb/kube-arangodb/tree/1.2.15) (2022-07-20)
 - (Bugfix) Ensure pod names not too long