+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - (Bugfix) Add Panic Handler
 - (Bugfix) Unify yaml packages
 - (Feature) Add new GRPC and HTTP API
+- (Feature) Add new API endpoints to allow getting and setting operator logging level
 
 ## [1.2.13](https://github.com/arangodb/kube-arangodb/tree/1.2.13) (2022-06-07)
 - (Bugfix) Fix arangosync members state inspection