+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - (Bugfix) Fix ClusterScaling integration
 - (Feature) Sensitive information protection
 - (Bugfix) Propagate SecurityContext to the ID Containers
+- (Feature) Propagate sidecars' ports to a member's service
 
 ## [1.2.15](https://github.com/arangodb/kube-arangodb/tree/1.2.15) (2022-07-20)
 - (Bugfix) Ensure pod names not too long