+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - (Bugfix) Propagate SecurityContext to the ID Containers
 - (Bugfix) Fix for enabling all features
 - (Feature) Propagate feature and predefined env variables to members
+- (Feature) Add progress action support
  
 ## [1.2.15](https://github.com/arangodb/kube-arangodb/tree/1.2.15) (2022-07-20)
 - (Bugfix) Ensure pod names not too long